+++ lib/features/chat/presentation/screens/chat_screen.dart
@@ -2,6 +2,8 @@ import 'package:chat_app/core/extensions/theme_extensions.dart';
 import 'package:chat_app/core/theme/app_colors.dart';
 import 'package:chat_app/core/utils/date_time_formatter.dart';
 import 'package:chat_app/core/widgets/app_scaffold.dart';
+import 'package:chat_app/features/chat/presentation/widgets/reply_preview.dart';
+import 'package:chat_app/features/chat/providers/reply_repository_provider.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
@@ -18,6 +20,7 @@ import 'package:chat_app/features/chat/providers/typing_provider.dart';
 import 'package:chat_app/features/chat/data/models/conversation.dart';
 import 'package:chat_app/features/chat/providers/conversation_message_provider.dart';
 import 'package:chat_app/features/chat/providers/conversation_provider.dart';
+import 'package:chat_app/features/chat/providers/reply_provider.dart';
 
 class ChatScreen extends ConsumerStatefulWidget {
   const ChatScreen({super.key, required this.conversation});
@@ -65,18 +68,23 @@ class _ChatScreenState extends ConsumerState<ChatScreen>
   //Future<void> _sendMessage(WidgetRef ref, String text) async {
   Future<void> _sendMessage(String text) async {
     final firebaseUser = FirebaseAuth.instance.currentUser;
-    if (firebaseUser == null) {
-      return;
-    }
-    final appUser = await ref.read(currentUserProvider.future);
-    if (appUser == null) {
-      return;
-    }
+    if (firebaseUser == null) return;
+
+    final appUser = ref.read(currentUserProvider).value;
+    if (appUser == null) return;
+
+    // 👇 Current reply (if any)
+    final reply = ref.read(replyProvider);
+
     final repository = ref.read(
-      conversationMessageRepositoryProvider(conversation.id),
+      conversationMessageRepositoryProvider(widget.conversation.id),
     );
+
+    final replyRepository = ref.read(replyRepositoryProvider);
+
     final document = repository.createMessageDocument();
-    final message = Message(
+
+    final draftMessage = Message(
       id: document.id,
       senderId: firebaseUser.uid,
       senderName: appUser.username,
@@ -90,29 +98,15 @@ class _ChatScreenState extends ConsumerState<ChatScreen>
       deletedAt: null,
     );
 
+    final message = replyRepository.attachReply(
+      draft: draftMessage,
+      reply: reply,
+    );
+
     await repository.sendMessage(message);
-    await ref
-        .read(conversationRepositoryProvider)
-        .updateConversationAfterMessage(
-          conversationId: conversation.id,
-          lastMessage: text,
-        );
-    await ref
-        .read(conversationRepositoryProvider)
-        .incrementUnread(
-          conversation: conversation,
-          senderId: firebaseUser.uid,
-        );
 
-    WidgetsBinding.instance.addPostFrameCallback((_) {
-      if (!_scrollController.hasClients) return;
-
-      _scrollController.animateTo(
-        _scrollController.position.maxScrollExtent,
-        duration: const Duration(milliseconds: 300),
-        curve: Curves.easeOut,
-      );
-    });
+    // Clear reply after successful send
+    ref.read(replyProvider.notifier).clear();
   }
 
   Future<void> _setOnline() async {
@@ -194,7 +188,7 @@ class _ChatScreenState extends ConsumerState<ChatScreen>
           children: [
             otherUserAsync.when(
               loading: () => Text('Loading...', style: context.titleText),
-              error: (_, __) => Text('Unknown User', style: context.titleText),
+              error: (_, _) => Text('Unknown User', style: context.titleText),
               data: (user) => Text(
                 user?.username ?? 'Unknown User',
                 style: context.titleText?.copyWith(
@@ -271,6 +265,8 @@ class _ChatScreenState extends ConsumerState<ChatScreen>
               );
             },
           ),
+
+          const ReplyPreview(),
 
           MessageInput(onSend: _sendMessage),
         ],